+++ oauth/metadata.go
@@ -20,6 +20,8 @@ type AuthorizationServerMetadataConfig struct {
 	AuthorizationEndpoint string
 	TokenEndpoint         string
 	RegistrationEndpoint  string
+	Resource              string
+	ResourceMetadataURL   string
 	ScopesSupported       []string
 }
 
@@ -36,8 +38,11 @@ type authorizationServerMetadata struct {
 	AuthorizationEndpoint         string   `json:"authorization_endpoint"`
 	TokenEndpoint                 string   `json:"token_endpoint"`
 	RegistrationEndpoint          string   `json:"registration_endpoint,omitempty"`
+	Resource                      string   `json:"resource,omitempty"`
+	ResourceMetadataURL           string   `json:"resource_metadata,omitempty"`
 	GrantTypesSupported           []string `json:"grant_types_supported"`
 	ResponseTypesSupported        []string `json:"response_types_supported"`
+	ResponseModesSupported        []string `json:"response_modes_supported"`
 	CodeChallengeMethodsSupported []string `json:"code_challenge_methods_supported"`
 	TokenEndpointAuthMethods      []string `json:"token_endpoint_auth_methods_supported"`
 	ScopesSupported               []string `json:"scopes_supported,omitempty"`
@@ -72,8 +77,11 @@ func AuthorizationServerMetadataHandler(cfg AuthorizationServerMetadataConfig) h
 			AuthorizationEndpoint:         cfg.AuthorizationEndpoint,
 			TokenEndpoint:                 cfg.TokenEndpoint,
 			RegistrationEndpoint:          cfg.RegistrationEndpoint,
+			Resource:                      cfg.Resource,
+			ResourceMetadataURL:           cfg.ResourceMetadataURL,
 			GrantTypesSupported:           []string{"authorization_code", "refresh_token"},
 			ResponseTypesSupported:        []string{"code"},
+			ResponseModesSupported:        []string{"query"},
 			CodeChallengeMethodsSupported: []string{"S256"},
 			TokenEndpointAuthMethods:      []string{authMethodNone, authMethodClientSecretBasic, authMethodClientSecretPost},
 			ScopesSupported:               append([]string{}, cfg.ScopesSupported...),